Huge pivot - simplified storage, save arrays
While initial version of the library is ~25% more efficient space-wise,
I found myself strugling with its complexity. I'm not ready for that
level of optimization at this early stage.

So, I'm changing approach to stare collection of graphs, where nodes and
labels are stored as is, while links represent from/to index pointers to
the arrays of nodes.